+++ self_improvement/engine.py
@@ -52,7 +52,7 @@ import importlib
 
 from db_router import GLOBAL_ROUTER, init_db_router
 
-from sandbox_settings import SandboxSettings
+from sandbox_settings import SandboxSettings, load_sandbox_settings
 from .init import (
     init_self_improvement,
     settings,
@@ -386,668 +386,18 @@ from ..self_improvement_policy import (
 )
 from ..pre_execution_roi_bot import PreExecutionROIBot, BuildTask, ROIResult
 from ..env_config import PRE_ROI_SCALE, PRE_ROI_BIAS, PRE_ROI_CAP
+from .learners import (
+    SynergyWeightLearner,
+    DQNSynergyLearner,
+    DoubleDQNSynergyLearner,
+    SACSynergyLearner,
+    TD3SynergyLearner,
+)
+from .dashboards import SynergyDashboard, load_synergy_history
 
 POLICY_STATE_LEN = 21
 
-
-class TorchReplayStrategy:
-    """Minimal DQN-style learner with replay buffer and configurable optimiser."""
-
-    def __init__(
-        self,
-        *,
-        net_factory: Callable[[int, int, Any], Any],
-        optimizer_cls: Any,
-        lr: float,
-        train_interval: int,
-        replay_size: int,
-        optimizer_kwargs: Mapping[str, Any] | None = None,
-        net_kwargs: Mapping[str, Any] | None = None,
-    ) -> None:
-        if sip_torch is None:  # pragma: no cover - torch not available
-            raise RuntimeError("PyTorch required for TorchReplayStrategy")
-        self.model = net_factory(7, 7, **(net_kwargs or {}))
-        self.optimizer = optimizer_cls(
-            self.model.parameters(), lr=lr, **(optimizer_kwargs or {})
-        )
-        self.loss_fn = sip_torch.nn.MSELoss()
-        self.train_interval = max(1, int(train_interval))
-        self.buffer: deque[tuple[Any, float]] = deque(maxlen=int(replay_size))
-        self.steps = 0
-        self.eval_loss = 0.0
-
-    # --------------------------------------------------------------
-    def update(
-        self,
-        state: Sequence[float],
-        reward: float,
-        extra: Mapping[str, float] | None = None,
-    ) -> list[float]:
-        self.steps += 1
-        if extra:
-            reward *= 1.0 + float(extra.get("avg_roi", 0.0))
-            reward *= 1.0 + float(extra.get("pass_rate", 0.0))
-        state_tensor = sip_torch.tensor(state, dtype=sip_torch.float32)
-        self.buffer.append((state_tensor, float(reward)))
-        if self.steps % self.train_interval == 0 and self.buffer:
-            states, rewards = zip(*self.buffer)
-            states_batch = sip_torch.stack(states)
-            rewards_batch = (
-                sip_torch.tensor(rewards, dtype=sip_torch.float32).unsqueeze(1)
-            )
-            targets = states_batch * rewards_batch
-            preds = self.model(states_batch)
-            loss = self.loss_fn(preds, targets)
-            self.optimizer.zero_grad()
-            loss.backward()
-            self.optimizer.step()
-            self.eval_loss = float(loss.item())
-            self.buffer.clear()
-        with sip_torch.no_grad():
-            q_vals = self.model(state_tensor).clamp(0.0, 10.0).tolist()
-        return [float(v) for v in q_vals]
-
-    # --------------------------------------------------------------
-    def save(self, base: str) -> bool:
-        try:
-            buf = io.BytesIO()
-            sip_torch.save(self.model.state_dict(), buf)
-            _atomic_write(Path(base + ".model.pt"), buf.getvalue(), binary=True)
-            buf = io.BytesIO()
-            sip_torch.save(self.optimizer.state_dict(), buf)
-            _atomic_write(Path(base + ".optim.pt"), buf.getvalue(), binary=True)
-            _atomic_write(
-                Path(base + ".replay.pkl"),
-                pickle.dumps(list(self.buffer)),
-                binary=True,
-            )
-        except Exception as exc:  # pragma: no cover - disk errors
-            logger.exception("failed to save strategy state: %s", exc)
-            return False
-        try:
-            sip_torch.load(base + ".model.pt")
-            sip_torch.load(base + ".optim.pt")
-        except Exception as exc:
-            logger.warning("failed to validate strategy save: %s", exc)
-            return False
-        return True
-
-    # --------------------------------------------------------------
-    def load(self, base: str) -> None:
-        try:
-            model_file = Path(base + ".model.pt")
-            optim_file = Path(base + ".optim.pt")
-            if model_file.exists():
-                self.model.load_state_dict(sip_torch.load(model_file))
-            if optim_file.exists():
-                self.optimizer.load_state_dict(sip_torch.load(optim_file))
-            replay_file = Path(base + ".replay.pkl")
-            if replay_file.exists():
-                with open(replay_file, "rb") as fh:
-                    items = pickle.load(fh)
-                self.buffer = deque(items, maxlen=self.buffer.maxlen)
-        except Exception as exc:  # pragma: no cover - disk errors
-            logger.warning("failed to load strategy state: %s", exc)
-
-
-class SynergyWeightLearner:
-    """Learner adjusting synergy weights using a simple RL policy."""
-
-    def __init__(
-        self,
-        path: Path | None = None,
-        lr: float | None = None,
-        *,
-        settings: SandboxSettings | None = None,
-        strategy_factory: Callable[..., Any] | None = None,
-        net_factory: Callable[..., Any] | None = None,
-        hyperparams: Mapping[str, Any] | None = None,
-    ) -> None:
-        """Create a learner.
-
-        Parameters
-        ----------
-        path:
-            Where weights and model state should be persisted.
-        lr:
-            Base learning rate for torch optimisers.
-        strategy_factory:
-            Callable returning an RL strategy instance.  Defaults to
-            :class:`ActorCriticStrategy`.
-        net_factory:
-            Optional factory to build the neural network used when PyTorch is
-            available.  Receives ``input_dim`` and ``output_dim`` along with any
-            ``net_kwargs`` from ``hyperparams``.
-        hyperparams:
-            Extra keyword arguments forwarded to ``strategy_factory`` and
-            optimisation components such as ``optimizer_cls``, ``optimizer_kwargs``
-            and ``net_kwargs``.
-        """
-
-        settings = settings or SandboxSettings()
-        if lr is None:
-            lr = float(getattr(settings, "synergy_weights_lr", 0.1))
-        self.train_interval = int(getattr(settings, "synergy_train_interval", 10))
-        self.replay_size = int(getattr(settings, "synergy_replay_size", 100))
-        self.path = Path(path) if path else Path(settings.synergy_weight_file)
-        self.lr = lr
-        self.weights = get_default_synergy_weights()
-        hp: Dict[str, Any] = dict(hyperparams or {})
-        strat_factory = strategy_factory or (lambda **kw: ActorCriticStrategy(**kw))
-        self.strategy = strat_factory(**hp)
-        self._state: tuple[float, ...] = (0.0,) * 7
-        self._steps = 0
-        self.eval_loss = 0.0
-        self.checkpoint_interval = int(
-            getattr(settings, "synergy_checkpoint_interval", 50)
-        )
-        self._save_count = 0
-        self.checkpoint_path = self.path.with_suffix(self.path.suffix + ".bak")
-        allow_py = bool(getattr(settings, "synergy_python_fallback", True))
-        max_replay = int(getattr(settings, "synergy_python_max_replay", 1000))
-        if sip_torch is None:
-            if not allow_py or self.replay_size > max_replay:
-                raise RuntimeError(
-                    "PyTorch is required for SynergyWeightLearner; pure-Python "
-                    "fallback is disabled or too slow",
-                )
-            logger.warning(
-                "PyTorch not installed - using %s strategy",
-                type(self.strategy).__name__,
-            )
-            self.buffer: deque[tuple[Any, float]] = deque(maxlen=self.replay_size)
-        else:
-            hidden = int(getattr(settings, "synergy_hidden_size", 32))
-            layers = int(getattr(settings, "synergy_layers", 1))
-            opt_name = str(getattr(settings, "synergy_optimizer", "adam")).lower()
-            opt_cls_default = {
-                "sgd": sip_torch.optim.SGD,
-                "adam": sip_torch.optim.Adam,
-            }.get(opt_name, sip_torch.optim.Adam)
-
-            def default_net(i: int, o: int, *, hidden_size: int = hidden, layers: int = layers):
-                modules: list[Any] = []
-                in_dim = i
-                for _ in range(max(0, layers)):
-                    modules.append(sip_torch.nn.Linear(in_dim, hidden_size))
-                    modules.append(sip_torch.nn.ReLU())
-                    in_dim = hidden_size
-                modules.append(sip_torch.nn.Linear(in_dim, o))
-                return sip_torch.nn.Sequential(*modules)
-
-            net_factory = net_factory or default_net
-            opt_kwargs = dict(hp.get("optimizer_kwargs", {}))
-            net_kwargs = dict(hp.get("net_kwargs", {}))
-            opt_cls = hp.get("optimizer_cls", opt_cls_default)
-            self.nn_strategy = TorchReplayStrategy(
-                net_factory=net_factory,
-                optimizer_cls=opt_cls,
-                lr=self.lr,
-                train_interval=self.train_interval,
-                replay_size=self.replay_size,
-                optimizer_kwargs=opt_kwargs,
-                net_kwargs=net_kwargs,
-            )
-            self.buffer = self.nn_strategy.buffer
-        self.load()
-
-    # ------------------------------------------------------------------
-    def load(self) -> None:
-        if not self.path or not self.path.exists():
-            return
-        data: dict[str, Any] | None = None
-        try:
-            with open(self.path, "r", encoding="utf-8") as fh:
-                data = json.load(fh)
-        except Exception as exc:
-            logger.warning("failed to load synergy weights %s: %s", self.path, exc)
-            if self.checkpoint_path.exists():
-                try:
-                    with open(self.checkpoint_path, "r", encoding="utf-8") as fh:
-                        data = json.load(fh)
-                        logger.warning(
-                            "recovered synergy weights from checkpoint %s",
-                            self.checkpoint_path,
-                        )
-                except Exception as exc2:
-                    logger.warning(
-                        "failed to load checkpoint %s: %s",
-                        self.checkpoint_path,
-                        exc2,
-                    )
-        if data is None:
-            self.weights = get_default_synergy_weights()
-            return
-
-        valid = isinstance(data, dict) and all(
-            k in data and isinstance(data[k], (int, float)) for k in self.weights
-        )
-        if not valid:
-            logger.warning(
-                "invalid synergy weight data in %s - using defaults", self.path
-            )
-            self.weights = get_default_synergy_weights()
-        else:
-            for k in self.weights:
-                self.weights[k] = float(data[k])
-        base = os.path.splitext(self.path)[0]
-        try:
-            pkl = base + ".policy.pkl"
-            if os.path.exists(pkl):
-                with open(pkl, "rb") as fh:
-                    self.strategy = pickle.load(fh)
-        except Exception as exc:
-            logger.exception("failed to load strategy pickle: %s", exc)
-        try:
-            replay = base + ".replay.pkl"
-            if os.path.exists(replay):
-                with open(replay, "rb") as fh:
-                    items = pickle.load(fh)
-                self.buffer = deque(items, maxlen=self.replay_size)
-        except Exception as exc:
-            logger.exception("failed to load replay buffer: %s", exc)
-        if sip_torch is not None and hasattr(self, "nn_strategy"):
-            self.nn_strategy.load(base)
-
-        logger.info("loaded synergy weights", extra=log_record(weights=self.weights))
-
-    # ------------------------------------------------------------------
-    def save(self) -> None:
-        if not self.path:
-            return
-        base = os.path.splitext(self.path)[0]
-        try:
-            _atomic_write(self.path, json.dumps(self.weights))
-            with open(self.path, "r", encoding="utf-8") as fh:
-                json.load(fh)
-        except Exception as exc:
-            logger.exception("failed to save synergy weights: %s", exc)
-            if self.checkpoint_path.exists():
-                try:
-                    shutil.copy(self.checkpoint_path, self.path)
-                except Exception as restore_exc:
-                    logger.exception(
-                        "failed to restore synergy weights from checkpoint %s",
-                        self.checkpoint_path,
-                        extra=log_record(checkpoint_path=str(self.checkpoint_path)),
-                    )
-                    raise restore_exc
-            return
-        self._save_count += 1
-        if self._save_count % self.checkpoint_interval == 0:
-            for attempt in range(3):
-                try:
-                    shutil.copy(self.path, self.checkpoint_path)
-                    break
-                except Exception:
-                    logger.exception(
-                        "failed to checkpoint synergy weights on attempt %s",
-                        attempt + 1,
-                    )
-                    if attempt == 2:
-                        raise
-                    time.sleep(0.1)
-        try:
-            pkl = Path(base + ".policy.pkl")
-            _atomic_write(pkl, pickle.dumps(self.strategy), binary=True)
-        except Exception as exc:
-            logger.exception("failed to save strategy pickle: %s", exc)
-        if sip_torch is not None and hasattr(self, "nn_strategy"):
-            ok = self.nn_strategy.save(base)
-            if not ok:
-                logger.warning("strategy persistence validation failed")
-        else:
-            try:
-                replay = Path(base + ".replay.pkl")
-                _atomic_write(replay, pickle.dumps(list(self.buffer)), binary=True)
-            except Exception as exc:
-                logger.exception("failed to save replay buffer: %s", exc)
-        try:
-            synergy_weight_updates_total.inc()
-        except Exception:
-            logger.exception("Unhandled exception in self_improvement")
-
-        logger.info("saved synergy weights", extra=log_record(weights=self.weights))
-
-    # ------------------------------------------------------------------
-    def update(
-        self,
-        roi_delta: float,
-        deltas: dict[str, float],
-        extra: dict[str, float] | None = None,
-    ) -> None:
-        names = [
-            "synergy_roi",
-            "synergy_efficiency",
-            "synergy_resilience",
-            "synergy_antifragility",
-            "synergy_reliability",
-            "synergy_maintainability",
-            "synergy_throughput",
-        ]
-        self._state = tuple(float(deltas.get(n, 0.0)) for n in names)
-        if sip_torch is not None and hasattr(self, "nn_strategy"):
-            q_vals_list = self.nn_strategy.update(self._state, float(roi_delta), extra)
-        else:
-            self.buffer.append((self._state, float(roi_delta)))
-            q_vals_list: list[float] = []
-            for idx, _ in enumerate(names):
-                reward = roi_delta * self._state[idx]
-                if extra:
-                    reward *= 1.0 + float(extra.get("avg_roi", 0.0))
-                    reward *= 1.0 + float(extra.get("pass_rate", 0.0))
-                q = self.strategy.update(
-                    {}, self._state, idx, reward, self._state, 1.0, 0.9
-                )
-                q_vals_list.append(float(q))
-            if hasattr(self.strategy, "predict"):
-                q_vals = self.strategy.predict(self._state)
-                q_vals_list = [float(v.item() if hasattr(v, "item") else v) for v in q_vals]
-        mapping = {
-            "roi": 0,
-            "efficiency": 1,
-            "resilience": 2,
-            "antifragility": 3,
-            "reliability": 4,
-            "maintainability": 5,
-            "throughput": 6,
-        }
-        for key, idx in mapping.items():
-            val = q_vals_list[idx]
-            self.weights[key] = max(0.0, min(float(val), 10.0))
-        logger.info(
-            "updated synergy weights",
-            extra=log_record(
-                weights=self.weights, state=self._state, roi_delta=roi_delta
-            ),
-        )
-        self.save()
-
-
-class DQNSynergyLearner(SynergyWeightLearner):
-    """Learner using a simple DQN to adapt synergy weights."""
-
-    def __init__(
-        self,
-        path: Path | None = None,
-        lr: float = 1e-3,
-        *,
-        strategy: str | None = None,
-        target_sync: int = 10,
-    ) -> None:
-        super().__init__(path, lr)
-        if hasattr(self, "nn_strategy"):
-            self.buffer = deque(maxlen=self.replay_size)
-            delattr(self, "nn_strategy")
-        if sip_torch is None:
-            logger.warning(
-                "PyTorch not available - DQNSynergyLearner falling back to DQN strategy"
-            )
-        name = (strategy or "dqn").lower()
-        self.target_sync = max(1, int(target_sync))
-        if (
-            name in {"double", "double_dqn", "double-dqn", "ddqn", "td3"}
-            and sip_torch is not None
-        ):
-            self.strategy = DoubleDQNStrategy(
-                action_dim=7, lr=lr, target_sync=target_sync
-            )
-            self.strategy_name = "td3" if name == "td3" else "double_dqn"
-        elif name in {
-            "policy",
-            "policy_gradient",
-            "actor_critic",
-            "actor-critic",
-            "sac",
-        }:
-            self.strategy = ActorCriticStrategy()
-            self.strategy_name = "sac" if name == "sac" else "policy_gradient"
-        else:
-            # prefer Double DQN when PyTorch is available
-            if sip_torch is not None:
-                self.strategy = DoubleDQNStrategy(
-                    action_dim=7, lr=lr, target_sync=target_sync
-                )
-                self.strategy_name = "double_dqn"
-            else:
-                self.strategy = DQNStrategy(action_dim=7, lr=lr)
-                self.strategy_name = "dqn"
-        self._state: tuple[float, ...] = (0.0,) * 7
-        self._steps = 0
-
-    # ------------------------------------------------------------------
-    def load(self) -> None:
-        super().load()
-        if not self.path:
-            return
-        base = os.path.splitext(self.path)[0]
-        if self.path and sip_torch is not None:
-            try:
-                if hasattr(self.strategy, "_ensure_model"):
-                    self.strategy._ensure_model(7)
-                pt = base + ".pt"
-                if os.path.exists(pt) and hasattr(self.strategy, "model"):
-                    state_dict = sip_torch.load(pt)
-                    assert self.strategy.model is not None
-                    self.strategy.model.load_state_dict(state_dict)
-                tgt = base + ".target.pt"
-                if os.path.exists(tgt) and hasattr(self.strategy, "target_model"):
-                    assert self.strategy.target_model is not None
-                    self.strategy.target_model.load_state_dict(sip_torch.load(tgt))
-            except Exception as exc:
-                logger.exception("failed to load DQN models: %s", exc)
-        try:
-            pkl = base + ".policy.pkl"
-            if os.path.exists(pkl):
-                with open(pkl, "rb") as fh:
-                    self.strategy = pickle.load(fh)
-        except Exception as exc:
-            logger.exception("failed to load strategy pickle: %s", exc)
-
-    # ------------------------------------------------------------------
-    def save(self) -> None:
-        super().save()
-        if not self.path:
-            return
-        base = os.path.splitext(self.path)[0]
-        if self.path and sip_torch is not None:
-            try:
-                if hasattr(self.strategy, "model") and self.strategy.model is not None:
-                    buf = io.BytesIO()
-                    sip_torch.save(self.strategy.model.state_dict(), buf)
-                    _atomic_write(Path(base + ".pt"), buf.getvalue(), binary=True)
-                if (
-                    hasattr(self.strategy, "target_model")
-                    and self.strategy.target_model is not None
-                ):
-                    buf = io.BytesIO()
-                    sip_torch.save(self.strategy.target_model.state_dict(), buf)
-                    _atomic_write(
-                        Path(base + ".target.pt"), buf.getvalue(), binary=True
-                    )
-            except Exception as exc:
-                logger.exception("failed to save DQN models: %s", exc)
-        try:
-            pkl = Path(base + ".policy.pkl")
-            _atomic_write(pkl, pickle.dumps(self.strategy), binary=True)
-        except Exception as exc:
-            logger.exception("failed to save strategy pickle: %s", exc)
-
-    # ------------------------------------------------------------------
-    def update(
-        self,
-        roi_delta: float,
-        deltas: dict[str, float],
-        extra: dict[str, float] | None = None,
-    ) -> None:
-        names = [
-            "synergy_roi",
-            "synergy_efficiency",
-            "synergy_resilience",
-            "synergy_antifragility",
-            "synergy_reliability",
-            "synergy_maintainability",
-            "synergy_throughput",
-        ]
-        self._state = tuple(float(deltas.get(n, 0.0)) for n in names)
-        q_vals_list: list[float] = []
-        for idx, name in enumerate(names):
-            reward = roi_delta * self._state[idx]
-            if extra:
-                reward *= 1.0 + float(extra.get("avg_roi", 0.0))
-                reward *= 1.0 + float(extra.get("pass_rate", 0.0))
-            q = self.strategy.update(
-                {}, self._state, idx, reward, self._state, 1.0, 0.9
-            )
-            q_vals_list.append(float(q))
-        if hasattr(self.strategy, "predict"):
-            q_vals = self.strategy.predict(self._state)
-            q_vals_list = [float(v.item() if hasattr(v, "item") else v) for v in q_vals]
-        mapping = {
-            "roi": 0,
-            "efficiency": 1,
-            "resilience": 2,
-            "antifragility": 3,
-            "reliability": 4,
-            "maintainability": 5,
-            "throughput": 6,
-        }
-        for key, idx in mapping.items():
-            val = q_vals_list[idx]
-            self.weights[key] = max(0.0, min(val, 10.0))
-        self._steps += 1
-        if (
-            self._steps % self.target_sync == 0
-            and hasattr(self.strategy, "target_model")
-            and hasattr(self.strategy, "model")
-            and getattr(self.strategy, "target_model") is not None
-            and getattr(self.strategy, "model") is not None
-        ):
-            try:
-                self.strategy.target_model.load_state_dict(
-                    self.strategy.model.state_dict()
-                )
-            except Exception as exc:
-                logger.exception("target model sync failed: %s", exc)
-        logger.info(
-            "updated synergy weights",
-            extra=log_record(
-                weights=self.weights, state=self._state, roi_delta=roi_delta
-            ),
-        )
-        self.save()
-
-
-class DoubleDQNSynergyLearner(DQNSynergyLearner):
-    """Synergy learner using a Double DQN strategy."""
-
-    def __init__(
-        self, path: Path | None = None, lr: float = 1e-3, *, target_sync: int = 10
-    ) -> None:
-        super().__init__(path, lr, strategy="double_dqn", target_sync=target_sync)
-
-
-class SACSynergyLearner(DQNSynergyLearner):
-    """Synergy learner using a simplified SAC strategy."""
-
-    def __init__(
-        self, path: Path | None = None, lr: float = 1e-3, *, target_sync: int = 10
-    ) -> None:
-        super().__init__(path, lr, strategy="sac", target_sync=target_sync)
-
-
-class TD3SynergyLearner(DQNSynergyLearner):
-    """Synergy learner using a simplified TD3 strategy."""
-
-    def __init__(
-        self, path: Path | None = None, lr: float = 1e-3, *, target_sync: int = 10
-    ) -> None:
-        super().__init__(path, lr, strategy="td3", target_sync=target_sync)
-
-
-def benchmark_workflow_variants(
-    workflow_id: int | str, variants: Mapping[str, Callable[[], bool]]
-) -> dict[str, tuple[EvaluationResult, float]]:
-    """Benchmark *variants* against the baseline workflow.
-
-    Parameters
-    ----------
-    workflow_id:
-        Identifier used for ROI tracking.
-    variants:
-        Mapping of variant identifiers to callables.  The mapping **must**
-        contain a ``"baseline"`` entry representing the current workflow.
-
-    Returns
-    -------
-    dict
-        Mapping of variant names to tuples of ``(EvaluationResult, roi_delta)``.
-    """
-
-    wf_id_str = str(workflow_id)
-    if "baseline" not in variants:
-        raise ValueError("variants must include a 'baseline' entry")
-
-    results_db = ROIResultsDB()
-
-    baseline_callable = variants["baseline"]
-    baseline_scorer = CompositeWorkflowScorer(results_db=results_db)
-    baseline_result = baseline_scorer.run(
-        baseline_callable, wf_id_str, run_id="baseline"
-    )
-    baseline_roi = baseline_result.roi_gain
-
-    results: dict[str, tuple[EvaluationResult, float]] = {
-        "baseline": (baseline_result, 0.0)
-    }
-
-    try:
-        wf_id_int = int(workflow_id)
-    except Exception:
-        wf_id_int = 0
-
-    for name, func in variants.items():
-        if name == "baseline":
-            continue
-        run_id = f"variant-{hash(name) & 0xffffffff:x}"
-        scorer = CompositeWorkflowScorer(results_db=results_db)
-        variant_result = scorer.run(func, wf_id_str, run_id=run_id)
-        roi_delta = variant_result.roi_gain - baseline_roi
-        results[name] = (variant_result, roi_delta)
-
-        results_db.log_module_delta(
-            wf_id_str,
-            run_id,
-            module=f"variant:{name}",
-            runtime=variant_result.runtime,
-            success_rate=variant_result.success_rate,
-            roi_delta=roi_delta,
-        )
-
-        try:
-            event_id = MutationLogger.log_mutation(
-                change=name,
-                reason="variant_evaluation",
-                trigger="benchmark_workflow_variants",
-                performance=0.0,
-                workflow_id=wf_id_int,
-                before_metric=baseline_roi,
-                after_metric=baseline_roi,
-            )
-            MutationLogger.record_mutation_outcome(
-                event_id,
-                after_metric=variant_result.roi_gain,
-                roi=roi_delta,
-                performance=roi_delta,
-            )
-        except Exception:  # pragma: no cover - logging shouldn't break flow
-            logging.getLogger(__name__).exception(
-                "mutation logging failed for variant %s", name
-            )
-
-    return results
-
-
+__all__ = ["SelfImprovementEngine"]
 class SelfImprovementEngine:
     """Run the automation pipeline on a configurable bot."""
 
@@ -7616,409 +6966,6 @@ class SelfImprovementEngine:
 
 
 from typing import Any, Callable, Optional, Type, Iterable, Sequence, Dict, List
-
-
-class ImprovementEngineRegistry:
-    """Register and run multiple :class:`SelfImprovementEngine` instances."""
-
-    def __init__(self) -> None:
-        self.engines: dict[str, SelfImprovementEngine] = {}
-
-    def register_engine(self, name: str, engine: SelfImprovementEngine) -> None:
-        """Add *engine* under *name*."""
-        self.engines[name] = engine
-
-    def unregister_engine(self, name: str) -> None:
-        """Remove the engine referenced by *name* if present."""
-        self.engines.pop(name, None)
-
-    def run_all_cycles(self, energy: int = 1) -> dict[str, AutomationResult]:
-        """Execute ``run_cycle`` on all registered engines.
-
-        Each returned :class:`AutomationResult` may include a
-        ``workflow_evolution`` field describing whether candidate
-        workflows were promoted or deemed stable.
-        """
-
-        results: dict[str, AutomationResult] = {}
-        for name, eng in self.engines.items():
-            if eng._should_trigger():
-                res = eng.run_cycle(energy=energy)
-                results[name] = res
-        return results
-
-    async def run_all_cycles_async(self, energy: int = 1) -> dict[str, AutomationResult]:
-        """Asynchronously execute ``run_cycle`` on all registered engines."""
-
-        async def _run(name: str, eng: SelfImprovementEngine):
-            if eng._should_trigger():
-                res = await asyncio.to_thread(eng.run_cycle, energy=energy)
-                return name, res
-            return None
-
-        tasks = [asyncio.create_task(_run(n, e)) for n, e in self.engines.items()]
-        results: dict[str, AutomationResult] = {}
-        for t in tasks:
-            out = await t
-            if out:
-                name, res = out
-                results[name] = res
-        return results
-
-    def schedule_all(
-        self, energy: int = 1, *, loop: asyncio.AbstractEventLoop | None = None
-    ) -> list[asyncio.Task]:
-        """Start schedules for all engines and return the created tasks."""
-        tasks: list[asyncio.Task] = []
-        for eng in self.engines.values():
-            tasks.append(eng.schedule(energy=energy, loop=loop))
-        return tasks
-
-    async def shutdown_all(self) -> None:
-        """Gracefully stop all running schedules."""
-        for eng in self.engines.values():
-            await eng.shutdown_schedule()
-
-    def autoscale(
-        self,
-        *,
-        capital_bot: CapitalManagementBot,
-        data_bot: DataBot,
-        factory: Callable[[str], SelfImprovementEngine],
-        max_engines: int = 5,
-        min_engines: int = 1,
-        create_energy: float = 0.8,
-        remove_energy: float = 0.3,
-        roi_threshold: float = 0.0,
-        cost_per_engine: float = 0.0,
-        approval_callback: Optional[Callable[[], bool]] = None,
-        max_instances: Optional[int] = None,
-    ) -> None:
-        """Dynamically create or remove engines based on ROI and resources."""
-        try:
-            energy = capital_bot.energy_score(
-                load=0.0,
-                success_rate=1.0,
-                deploy_eff=1.0,
-                failure_rate=0.0,
-            )
-        except Exception as exc:
-            self.logger.exception("autoscale energy check failed: %s", exc)
-            energy = 0.0
-        try:
-            trend = data_bot.long_term_roi_trend(limit=200)
-        except Exception as exc:
-            self.logger.exception("autoscale trend fetch failed: %s", exc)
-            trend = 0.0
-        if not capital_bot.check_budget():
-            return
-        if max_instances is not None and len(self.engines) >= max_instances:
-            return
-        projected_roi = trend - cost_per_engine
-        if (
-            energy >= create_energy
-            and trend > roi_threshold
-            and projected_roi > 0.0
-            and len(self.engines) < max_engines
-        ):
-            if approval_callback and not approval_callback():
-                return
-            name = f"engine{len(self.engines)}"
-            self.register_engine(name, factory(name))
-            return
-        if (
-            energy <= remove_energy or trend <= roi_threshold or projected_roi <= 0.0
-        ) and len(self.engines) > min_engines:
-            name = next(iter(self.engines))
-            self.unregister_engine(name)
-
-
-__all__ = [
-    "SelfImprovementEngine",
-    "ImprovementEngineRegistry",
-    "benchmark_workflow_variants",
-    "auto_x",
-    "SynergyDashboard",
-    "load_synergy_history",
-    "synergy_stats",
-    "synergy_ma",
-]
-
-
-def auto_x(
-    engines: list[SelfImprovementEngine] | None = None,
-    *,
-    energy: int = 1,
-) -> dict[str, AutomationResult]:
-    """Convenience helper to run a self‑improvement cycle."""
-    registry = ImprovementEngineRegistry()
-    if engines:
-        for idx, eng in enumerate(engines):
-            registry.register_engine(f"engine{idx}", eng)
-    else:
-        registry.register_engine("default", SelfImprovementEngine())
-    results = registry.run_all_cycles(energy=energy)
-    return results
-
-
-def load_synergy_history(path: str | Path) -> list[dict[str, float]]:
-    """Return synergy history entries from ``path`` SQLite database."""
-    p = Path(path)
-    if not p.exists():
-        return []
-    try:
-        rows = (
-            router.get_connection("synergy_history")
-            .execute("SELECT entry FROM synergy_history ORDER BY id")
-            .fetchall()
-        )
-        hist: list[dict[str, float]] = []
-        for (text,) in rows:
-            data = json.loads(text)
-            if isinstance(data, dict):
-                hist.append({str(k): float(v) for k, v in data.items()})
-        return hist
-    except Exception:
-        return []
-
-
-def synergy_stats(history: list[dict[str, float]]) -> dict[str, dict[str, float]]:
-    """Return average and variance for each synergy metric."""
-    metrics: dict[str, list[float]] = {}
-    for entry in history:
-        for k, v in entry.items():
-            metrics.setdefault(str(k), []).append(float(v))
-    stats: dict[str, dict[str, float]] = {}
-    for name, vals in metrics.items():
-        arr = np.array(vals, dtype=float)
-        stats[name] = {
-            "average": float(arr.mean()) if arr.size else 0.0,
-            "variance": float(arr.var()) if arr.size else 0.0,
-        }
-    return stats
-
-
-def synergy_ma(
-    history: list[dict[str, float]], window: int = 5
-) -> list[dict[str, float]]:
-    """Return rolling averages over ``window`` samples for each metric."""
-    if window < 1:
-        raise ValueError("window must be positive")
-    metrics = sorted({k for d in history for k in d})
-    ma_history: list[dict[str, float]] = []
-    for idx in range(len(history)):
-        ma_entry: dict[str, float] = {}
-        start = max(0, idx + 1 - window)
-        for name in metrics:
-            vals = [history[j].get(name, 0.0) for j in range(start, idx + 1)]
-            arr = np.array(vals, dtype=float)
-            ma_entry[name] = float(arr.mean()) if arr.size else 0.0
-        ma_history.append(ma_entry)
-    return ma_history
-
-
-class SynergyDashboard:
-    """Expose synergy metrics via a small Flask app.
-
-    Parameters
-    ----------
-    history_file:
-        Path to the SQLite history database.
-    ma_window:
-        Window size for moving averages.
-    exporter_host/exporter_port:
-        When set, fetch metrics from a Prometheus exporter instead of the
-        history file.
-    refresh_interval:
-        How often to poll the exporter.
-    max_history:
-        Maximum number of history entries to keep in memory when using an
-        exporter. ``None`` disables trimming.
-    """
-
-    def __init__(
-        self,
-        history_file: str | Path = "synergy_history.db",
-        *,
-        ma_window: int = 5,
-        exporter_host: str | None = None,
-        exporter_port: int = 8003,
-        refresh_interval: float = 5.0,
-        max_history: int | None = None,
-    ) -> None:
-        from flask import Flask, jsonify  # type: ignore
-
-        self.logger = get_logger(self.__class__.__name__)
-        self.history_file = Path(history_file)
-        self.ma_window = ma_window
-        self.exporter_host = exporter_host
-        self.exporter_port = exporter_port
-        self.refresh_interval = float(refresh_interval)
-        self.max_history = max_history
-        self._history: list[dict[str, float]] = []
-        self._last_metrics: dict[str, float] = {}
-        self._stop = threading.Event()
-        self._thread: threading.Thread | None = None
-        self.jsonify = jsonify
-        self.app = Flask(__name__)
-        self.app.add_url_rule("/", "index", self.index)
-        self.app.add_url_rule("/stats", "stats", self.stats)
-        self.app.add_url_rule("/plot.png", "plot", self.plot)
-        self.app.add_url_rule("/history", "history", self.history)
-
-        if self.exporter_host:
-            self._thread = threading.Thread(target=self._update_loop, daemon=True)
-            self._thread.start()
-
-    def _load(self) -> list[dict[str, float]]:
-        if self.exporter_host:
-            return list(self._history)
-        return load_synergy_history(self.history_file)
-
-    # --------------------------------------------------------------
-    def _parse_metrics(self, text: str) -> dict[str, float]:
-        metrics: dict[str, float] = {}
-        for line in text.splitlines():
-            if line.startswith("#"):
-                continue
-            parts = line.strip().split()
-            if len(parts) != 2:
-                continue
-            name, value = parts
-            if name.startswith("synergy_"):
-                try:
-                    metrics[name] = float(value)
-                except ValueError:
-                    continue
-        return metrics
-
-    def _fetch_exporter_metrics(self) -> dict[str, float]:
-        import requests  # type: ignore
-
-        url = f"http://{self.exporter_host}:{self.exporter_port}/metrics"
-        try:
-            resp = requests.get(url, timeout=1.0)
-            if resp.status_code != 200:
-                raise RuntimeError(f"status {resp.status_code}")
-            metrics = self._parse_metrics(resp.text)
-            if metrics:
-                self._last_metrics = metrics
-            return metrics
-        except Exception as exc:  # pragma: no cover - runtime issues
-            self.logger.warning("failed to fetch metrics from %s: %s", url, exc)
-            return dict(self._last_metrics)
-
-    def _update_loop(self) -> None:
-        while not self._stop.is_set():
-            vals = self._fetch_exporter_metrics()
-            if vals:
-                self._history.append(vals)
-                if self.max_history and len(self._history) > self.max_history:
-                    self._history = self._history[-self.max_history :]
-            self._stop.wait(self.refresh_interval)
-
-    def stop(self) -> None:
-        self._stop.set()
-        if self._thread:
-            self._thread.join(timeout=1.0)
-
-    def index(self) -> tuple[str, int]:
-        return "Synergy dashboard running. Access /stats", 200
-
-    def stats(self) -> tuple[str, int]:
-        hist = self._load()
-        ma_hist = synergy_ma(hist, self.ma_window)
-        data = {
-            "stats": synergy_stats(hist),
-            "latest": hist[-1] if hist else {},
-            "rolling_average": ma_hist[-1] if ma_hist else {},
-        }
-        return self.jsonify(data), 200
-
-    def history(self) -> tuple[list[dict[str, float]], int]:
-        return self.jsonify(self._load()), 200
-
-    def plot(self) -> tuple[bytes, int, dict[str, str]]:
-        try:
-            import matplotlib.pyplot as plt  # type: ignore
-        except Exception:
-            return b"", 200, {"Content-Type": "image/png"}
-
-        history = self._load()
-        ma_history = synergy_ma(history, self.ma_window)
-        metrics = sorted({k for d in history for k in d})
-        labels = list(range(len(history)))
-        fig, ax = plt.subplots()
-        for name in metrics:
-            vals = [d.get(name, 0.0) for d in history]
-            ax.plot(labels, vals, label=name)
-            ma_vals = [d.get(name, 0.0) for d in ma_history]
-            ax.plot(labels, ma_vals, label=f"{name}_ma", linestyle="--")
-        if metrics:
-            ax.legend()
-        ax.set_xlabel("iteration")
-        ax.set_ylabel("value")
-        fig.tight_layout()
-        from ..import BytesIO
-
-        buf = BytesIO()
-        fig.savefig(buf, format="png")
-        plt.close(fig)
-        buf.seek(0)
-        return buf.getvalue(), 200, {"Content-Type": "image/png"}
-
-    def run(
-        self,
-        host: str = "0.0.0.0",
-        port: int = 5001,
-        *,
-        wsgi: str = "flask",
-    ) -> None:
-        """Run the dashboard using the selected WSGI/ASGI server."""
-        with contextlib.closing(socket.socket()) as sock:
-            try:
-                sock.bind((host, port))
-            except OSError:
-                self.logger.error("port %d in use", port)
-                raise
-
-        server = wsgi.lower()
-        if server == "flask":
-            self.app.run(host=host, port=port)
-            return
-
-        if server == "gunicorn":
-            try:
-                from gunicorn.app.base import BaseApplication  # type: ignore
-            except Exception as exc:  # pragma: no cover - optional dependency
-                raise RuntimeError("gunicorn is required for --wsgi gunicorn") from exc
-
-            class _GunicornApp(BaseApplication):
-                def __init__(self, app):
-                    self.application = app
-                    super().__init__()
-
-                def load_config(self):  # pragma: no cover - runtime setup
-                    self.cfg.set("bind", f"{host}:{port}")
-                    self.cfg.set("workers", 1)
-
-                def load(self):  # pragma: no cover - runtime setup
-                    return self.application
-
-            _GunicornApp(self.app).run()
-            return
-
-        if server == "uvicorn":
-            try:
-                import uvicorn  # type: ignore
-                from starlette.middleware.wsgi import WSGIMiddleware
-            except Exception as exc:  # pragma: no cover - optional dependency
-                raise RuntimeError("uvicorn is required for --wsgi uvicorn") from exc
-
-            uvicorn.run(WSGIMiddleware(self.app), host=host, port=port, workers=1)
-            return
-
-        raise ValueError(f"unknown wsgi server: {wsgi}")
 
 
 def cli(argv: list[str] | None = None) -> None:
